wip: eliminate GPT image smells
